Move map from home to gardens page.
