+++ components/views/Home.js
@@ -1,14 +1,6 @@
 import html from "html-literal";
-
-const kelvinToFahrenheit = (kelvinTemp) =>
-  Math.round((kelvinTemp - 273.15) * (9 / 5) + 32);
+import gardenBackground from "/assets/img/jed-owen-1JgUGDdcWnM-unsplash.jpg";
 
 export default (st) => html`
-  <h3>
-    Temperature in ${st.weather.city} is
-    ${kelvinToFahrenheit(st.weather.temp)}F. It feels like
-    ${kelvinToFahrenheit(st.weather.feelsLike)}F.
-  </h3>
-
-  <div id="map"></div>
+  <img src="${gardenBackground}" />
 `;
