+++ store/Gardens.js
@@ -1,4 +1,5 @@
 export default {
-    view: "Gardens",
-    gardens: []
+  view: "Gardens",
+  gardens: [],
+  // index.js open weather api creates a weather object here.
 };
