+++ store/Home.js
@@ -1,4 +1,3 @@
 export default {
     view: "Home"
-    // index.js open weather api creates a weather object here.
 };
